Revert "Move common types between server and client into "common" (#17)"
This reverts commit 29446aca3bbf3eed4518b132f9533e19a56b194f.
