+++ app/src/interfaces/bus.ts
@@ -10,3 +10,4 @@ export interface Bus {
     routeId: string;
   };
 }
+
