--- common/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "real-time-bus-common-types",
-  "private": true
-}
