--- common/route.ts
@@ -1,6 +1,0 @@
-export interface Route {
-  route_id: string;
-  route_short_name?: string;
-  route_long_name?: string;
-  route_type?: string;
-}
